wc-api: Rename methods
This renames apirequest-methods to just methods.
